requirements.txt
# general modules
django
python-decouple
djangorestframework

# third party modules
markdown       
django-filter  
drf-yasg[validation]
djangorestframework-simplejwt
django-simple-captcha

# email third party modules
django-mail-templated

# depolyment modules
